+++ tests/Benchmark/Program.cs
@@ -7,10 +7,10 @@ static class Program
     static void Main(string[] args)
     {
         BenchmarkRunner.Run<HighLoadBenchmarks>();
-        BenchmarkRunner.Run<InitializingCurrencyBenchmarks>();
-        BenchmarkRunner.Run<InitializingMoneyBenchmarks>();
-        BenchmarkRunner.Run<MoneyOperationsBenchmarks>();
-        BenchmarkRunner.Run<MoneyFormattingBenchmarks>();
-        BenchmarkRunner.Run<MoneyParsingBenchmarks>();
+        // BenchmarkRunner.Run<InitializingCurrencyBenchmarks>();
+        // BenchmarkRunner.Run<InitializingMoneyBenchmarks>();
+        // BenchmarkRunner.Run<MoneyOperationsBenchmarks>();
+        // BenchmarkRunner.Run<MoneyFormattingBenchmarks>();
+        // BenchmarkRunner.Run<MoneyParsingBenchmarks>();
     }
 }